apps/web/requirements.txt
Flask==3.0.3
rembg==2.0.57
numpy==1.26.3
pillow==9.5.0
uvicorn[standard]==0.30.1
pydantic==2.6.0